Small fix in env.rc
We have to put the tools path before the user path because we
must use our compiled gcc and not the system version.
But this is not ideal since that way we cannot have them always
in our path.

The ideal situation would be to use exactly the same gcc version that
is installed system-wide, but gcc 9 seems to break code coverage in ghdl